Moving some structs to a separate file
